subscriber redirect

+++ accounts/models.py
@@ -46,7 +46,7 @@ class StripePlan(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE, related_name='plans')
     plan_id = models.CharField(max_length=100)
     name = models.CharField(max_length=100)
-    amount = models.IntegerField()
+    amount = models.DecimalField(max_digits=9, decimal_places=2)
     currency = models.CharField(max_length=3)
     interval = models.CharField(max_length=10)
